Title: 버스를 기다리며, Time: 30ms, Memory: 1MB, Status: Wrong Answer - CodeTree

+++ 250218/버스를 기다리며/waiting-for-the-bus.cpp
@@ -3,25 +3,25 @@
 
 using namespace std;
 
-int n, m, c;
-int arr[100001];
-int l = 0, r = 1e9, ans = 0;
+long long n, m, c;
+long long arr[100001];
+long long l = 0, r = 1e9, ans = 0;
 
 void Input() {
 	cin >> n >> m >> c;
 
-	for (int i = 0; i < n; i++) {
+	for (long long i = 0; i < n; i++) {
 		cin >> arr[i];
 	}
 
 	sort(arr, arr + n);
 }
 
-bool Calc(int mid) {
-	int bus_cnt = 1;
-	int last_person = arr[0];
+bool Calc(long long mid) {
+	long long bus_cnt = 1;
+	long long last_person = arr[0];
 
-	for (int i = 1; i < n; i++) {
+	for (long long i = 1; i < n; i++) {
 		if (arr[i] - last_person <= mid && bus_cnt < c) {
 			continue;
 		}
@@ -41,7 +41,7 @@ int main() {
 	Input();
 
 	while (l <= r) {
-		int mid = (l + r) / 2;
+		long long mid = (l + r) / 2;
 
 		if (Calc(mid)) {
 			r = mid - 1;
